Move notify function to utils
