--- src/common/Notification/Notification.jsx
@@ -1,21 +1,0 @@
-// import React from 'react';
-// import { Space, notification } from 'antd';
-
-// const close = () => {
-//   console.log(
-//     'Notification was closed. Either the close button was clicked or duration time elapsed.',
-//   );
-// };
-// const Notification = () => {
-//   const [api] = notification.useNotification();
-//   const key = `open${Date.now()}`;
-//   api.open({
-//     message: 'Notification Title',
-//     description:
-//       'A function will be be called after the notification is closed (automatically after the "duration" time of manually).',
-//     key,
-//     onClose: close,
-// });
-// };
-
-// export default Notification;
