+++ src/pages/ManageUsers/ManageUsersPage.module 2.css
@@ -3,8 +3,6 @@
   justify-content: center;
 }
 
-.page{
-  width:75%;
+.page {
+  width: 75%;
 }
-
-
